Compiles. I refactored the program massively and came across a few more improvements that could be made. I'm starting to think the Stats arraylist shouldn't be global so that functions can be reused more effectively. I get that it can be rewritten each iteration but in some parts of the program it's being used in a context that doesn't suit the data gotten from the interactive part... Maybe I'm wrong? and I think the program works as intended now so I guess no need to fix something that isn't broken.
What is the heuristic for commit messages? does the above belong somewhere else? Anyway the refactoring I did was mainly decoupling the logging and the generating. Now the program generates the data first and logs it afterwards. Actually it logs each day instantly and the totals afterwards (as was intended in the specification). I did some polishing as well and put in explanations / speculations on a whole bunch of code.